Decompose scan and findings modules
