+++ src/agents_shipgate/cli/_register_baseline.py
@@ -6,7 +6,8 @@ from pathlib import Path
 import typer
 
 from agents_shipgate.checks.baseline_integrity import has_hash_mismatch
-from agents_shipgate.cli.scan import _resolve_audit_log_path, run_scan
+from agents_shipgate.cli.scan.orchestrator import run_scan
+from agents_shipgate.cli.scan.path_helpers import _resolve_audit_log_path
 from agents_shipgate.config.loader import load_manifest
 from agents_shipgate.core.baseline import verify_baseline, write_baseline
 from agents_shipgate.core.errors import AgentsShipgateError, ConfigError, InputParseError
