+++ src/agents_shipgate/cli/scan/__init__.py
@@ -1,0 +1,119 @@
+from __future__ import annotations
+
+from agents_shipgate.config.loader import load_manifest, load_manifest_with_positions
+from agents_shipgate.inputs.protocol import REGISTRY
+
+from .agent_builder import _build_agent, _first_adk_instruction_preview
+from .decision import _run_checks_and_decide
+from .diffs import _load_diff_references
+from .final_report import _build_final_report
+from .inputs import _load_inputs
+from .inspect import inspect_sources
+from .models import (
+    _ChecksDecision,
+    _DiffReferences,
+    _LoadedInputs,
+    _OutputPlan,
+    _ResolvedManifest,
+    _SanitizedSurfaces,
+    _ToolsAndAgent,
+)
+from .orchestrator import run_scan
+from .output_helpers import (
+    PACKET_FORMAT_NAMES,
+    _planned_generated_paths,
+    _resolve_packet_format_set,
+    _write_packet,
+    _write_reports,
+)
+from .output_planning import _plan_outputs
+from .patching import _attach_patches, _check_metadata_lookup
+from .path_helpers import (
+    _default_baseline_status,
+    _relative_display_path,
+    _resolve_audit_log_path,
+)
+from .prepare import _prepare_scan
+from .run_identity import _run_id
+from .sanitization import _sanitize_for_output
+from .source_loading import (
+    _absorb,
+    _artifact_warnings,
+    _flatten_and_deduplicate_tools,
+    _invoke_per_source_adapter,
+    _load_sources,
+    _manifest_placeholder_warnings,
+    _merge_duplicate_tool_metadata,
+    _merge_string_values,
+    _resolve_source_paths,
+    _risk_hint_key,
+    _source_priority,
+    _tool_source_index,
+)
+from .surface_sanitization import (
+    _build_public_action_surface_facts,
+    _disambiguate_public_action_ids,
+    _frameworks_surface,
+    _refresh_public_action_hashes,
+    _sanitize_codex_plugin_surface,
+    _sanitize_diff_reference,
+    _sanitize_existing_action_surface_facts,
+)
+from .tools_agent import _build_tools_and_agent
+from .writing import _write_outputs
+
+__all__ = [
+    "PACKET_FORMAT_NAMES",
+    "REGISTRY",
+    "_ChecksDecision",
+    "_DiffReferences",
+    "_LoadedInputs",
+    "_OutputPlan",
+    "_ResolvedManifest",
+    "_SanitizedSurfaces",
+    "_ToolsAndAgent",
+    "_absorb",
+    "_artifact_warnings",
+    "_attach_patches",
+    "_build_agent",
+    "_build_final_report",
+    "_build_public_action_surface_facts",
+    "_build_tools_and_agent",
+    "_check_metadata_lookup",
+    "_default_baseline_status",
+    "_disambiguate_public_action_ids",
+    "_first_adk_instruction_preview",
+    "_flatten_and_deduplicate_tools",
+    "_frameworks_surface",
+    "_invoke_per_source_adapter",
+    "_load_diff_references",
+    "_load_inputs",
+    "_load_sources",
+    "_manifest_placeholder_warnings",
+    "_merge_duplicate_tool_metadata",
+    "_merge_string_values",
+    "_plan_outputs",
+    "_planned_generated_paths",
+    "_prepare_scan",
+    "_refresh_public_action_hashes",
+    "_relative_display_path",
+    "_resolve_audit_log_path",
+    "_resolve_packet_format_set",
+    "_resolve_source_paths",
+    "_risk_hint_key",
+    "_run_checks_and_decide",
+    "_run_id",
+    "_sanitize_codex_plugin_surface",
+    "_sanitize_diff_reference",
+    "_sanitize_existing_action_surface_facts",
+    "_sanitize_for_output",
+    "_source_priority",
+    "_tool_source_index",
+    "_write_outputs",
+    "_write_packet",
+    "_write_reports",
+    "inspect_sources",
+    "load_manifest",
+    "load_manifest_with_positions",
+    "run_scan",
+]
